Refactor live and mock runners into one and add logging

+++ src/live_logic/parameters.py
@@ -11,7 +11,7 @@ class Parameters:
 class LiveParameters(Parameters):
     def __init__(self, update_period: timedelta,
                  trade_amount: int,
-                 sleep_time: int):
+                 sleep_time: float):
         self.update_period = update_period
         self.trade_amount = trade_amount
         self.sleep_time = sleep_time
